+++ src/test/java/qna/domain/AnswerRepositoryTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.fixture.UserTestFixture;
 
 import java.util.Arrays;
 import java.util.List;
