+++ src/test/java/qna/fixture/QuestionTestFixture.java
@@ -1,4 +1,6 @@
-package qna.domain;
+package qna.fixture;
+
+import qna.domain.Question;
 
 public class QuestionTestFixture {
     public static final Question Q1 = new Question(1L, UserTestFixture.JAVAJIGI, "title1", "contents1").writeBy(UserTestFixture.JAVAJIGI);
